exercise

+++ exam_tasks_24/system_split2_16820/project/software/express_software.py
@@ -4,5 +4,5 @@ from project.software.software import Software
 class ExpressSoftware(Software):
 
     def __init__(self, name: str, capacity_consumption, memory_consumption: int):
-        super().__init__(name, "Express", capacity_consumption, int(memory_consumption*2))
+        super().__init__(name, "Express", capacity_consumption, memory_consumption*2)
